backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic==2.5.0
uvicorn[standard]==0.24.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
PyJWT==2.8.0
